% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework5_sol.R
\name{backtrack_descent_sol}
\alias{backtrack_descent_sol}
\title{Backtracking for steepest descent}
\usage{
backtrack_descent_sol(fx, x, t, df, d, alpha = 0.5, beta = 0.9)
}
\arguments{
\item{fx}{handle to function that returns objective function values}

\item{x}{current parameter estimate}

\item{t}{current step-size}

\item{df}{the value of the gradient of objective function evaluated at the current x}

\item{d}{descent direction vector}

\item{alpha}{the backtracking parameter}

\item{beta}{the decrementing multiplier}
}
\description{
Backtracking for steepest descent
}
\examples{
set.seed(1245)
p <- 10
fx <- function(x) {
   return(0.5*sum(c(x) ** 2))
}
x0 <- matrix(rnorm(p), ncol=1)
df <- x0
d <- matrix(0, p, ncol=1)
ix <- 1:floor(p/2)
d[ix] <- - df[ix]
t <- 10
backtrack_descent_sol(fx=fx, x=x0, t=t, df=df, d=d)
}
