% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_k_banded.R
\name{check_k_banded}
\alias{check_k_banded}
\title{Check Matrix is k-banded}
\usage{
check_k_banded(A, k)
}
\arguments{
\item{A}{The input matrix}

\item{k}{bandwidth parameter}
}
\description{
\code{check_k_banded} returns a Boolean variable indicating whether or
not a matrix is k-banded.
}
