% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework5_sol.R
\name{dd_hsvm}
\alias{dd_hsvm}
\title{Pseudo-Hessian for HSVM}
\usage{
dd_hsvm(y, X, beta, delta, lambda)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{beta}{regression coefficient vector}

\item{delta}{shape parameter}

\item{lambda}{regularization parameter}
}
\description{
Pseudo-Hessian for HSVM
}
