% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework5_sol.R
\name{dphi}
\alias{dphi}
\title{Huberized Hinge Loss Derivative}
\usage{
dphi(t, delta)
}
\arguments{
\item{t}{input vector}

\item{delta}{shape parameter}
}
\description{
Huberized Hinge Loss Derivative
}
\examples{
n <- 1e3
delta <- 2
t <- seq(-2, 2, length.out=n)
y <- dphi(t, delta)
plot(t, y, type='l')
}
