% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_hsvm.R, R/homework5_sol.R
\name{fx_hsvm}
\alias{fx_hsvm}
\alias{fx_hsvm}
\title{Objective Function for HSVM}
\usage{
fx_hsvm(y, X, beta, delta, lambda = 1e-04)

fx_hsvm(y, X, beta, delta, lambda = 1e-04)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{beta}{regression coefficient vector}

\item{delta}{shape parameter}

\item{lambda}{regularization parameter}

\item{y}{binary response}

\item{X}{design matrix}

\item{beta}{regression coefficient vector}

\item{delta}{shape parameter}

\item{lambda}{regularization parameter}
}
\description{
Objective Function for HSVM

Objective Function for HSVM
}
