% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient Descent}
\usage{
gradient_descent(y, X, b0, delta, t = NULL, lambda = 1e-04,
  max_iter = 100, tol = 0.001)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{b0}{initial parameter estimate}

\item{delta}{shape parameter}

\item{t}{step-size}

\item{lambda}{regularization parameter}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Gradient Descent
}
