% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent_fixed.R
\name{gradient_descent_fixed}
\alias{gradient_descent_fixed}
\title{Gradient Descent (Fixed Step-Size)}
\usage{
gradient_descent_fixed(y, X, b0, t, delta, lambda = 1e-04, max_iter = 100,
  tol = 0.001)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{b0}{initial parameter estimate}

\item{t}{step-size}

\item{delta}{shape parameter}

\item{lambda}{regularization parameter}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Gradient Descent (Fixed Step-Size)
}
