% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSVM_2nd_order.R
\name{hsvm_newton}
\alias{hsvm_newton}
\title{Damped Newton's Method for Fitting HSVM}
\usage{
hsvm_newton(X, y, b, delta, lambda = 1e-04, max_iter = 100, tol = 0.001,
  naive = 1)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Binary response}

\item{b}{Initial regression coefficient vector}

\item{delta}{HSVM parameter}

\item{lambda}{regularization parameter}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}

\item{naive}{which method to use}
}
\description{
Damped Newton's Method for Fitting HSVM
}
