% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isweep.R
\name{isweep}
\alias{isweep}
\title{Inverse Sweep}
\usage{
isweep(A, k = NULL)
}
\arguments{
\item{A}{The input symmetric matrix}

\item{k}{Diagonal index entry set on which to sweep}
}
\description{
Inverse Sweep
}
