% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isweep_k.R
\name{isweep_k}
\alias{isweep_k}
\title{Inverse Sweep k}
\usage{
isweep_k(A, k)
}
\arguments{
\item{A}{The input symmetric matrix}

\item{k}{Diagonal index on which to sweep}
}
\description{
\code{isweep_k} applies the inverse sweep operator to a symmetric matrix  on the kth diagonal entry if it is possible.
}
