% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSVM_2nd_order.R
\name{newton_step_naive}
\alias{newton_step_naive}
\title{Compute quasi-Newton Step (Naive) for HSVM}
\usage{
newton_step_naive(X, y, b, g, delta, lambda = 1e-04)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Binary response vector}

\item{b}{Regression vector}

\item{g}{Gradient vector}

\item{delta}{HSVM parameter}

\item{lambda}{Regulaization parameter}
}
\description{
Compute quasi-Newton Step (Naive) for HSVM
}
