% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework5_sol.R
\name{phi}
\alias{phi}
\title{Huberized Hinge Loss}
\usage{
phi(t, delta)
}
\arguments{
\item{t}{input vector}

\item{delta}{shape parameter}
}
\description{
Huberized Hinge Loss
}
\examples{
n <- 1e3
delta <- 2
t <- seq(-2, 2, length.out=n)
y <- phi(t, delta)
plot(t, y, type='l')
}
