% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completion_sol.R
\name{project_omega_sol}
\alias{project_omega_sol}
\title{Project matrix onto observed index set}
\usage{
project_omega_sol(X, Omega)
}
\arguments{
\item{X}{input matrix}

\item{Omega}{index set}
}
\description{
Project matrix onto observed index set
}
\examples{
set.seed(12345)
n <- 10; p <- 20
X <- matrix(rnorm(n*p), n, p)
numel <- n*p
Omega <- sample(1:numel, floor(0.5*numel), replace=FALSE)
PX <- project_omega_sol(X, Omega)
}
