% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{prox_grad}
\alias{prox_grad}
\title{Proximal Gradient Descent}
\usage{
prox_grad(X, U, Omega, lambda, t, max_iter = 100, tol = 0.001)
}
\arguments{
\item{X}{input matrix}

\item{U}{current guess}

\item{Omega}{index set}

\item{lambda}{regularization parameter}

\item{t}{step-size parameter}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Proximal Gradient Descent
}
\examples{
set.seed(12345)
n <- 10; p <- 20
X <- matrix(rnorm(n*p), n, p)
numel <- n*p
Omega <- sample(1:numel, floor(0.5*numel), replace=FALSE)
U <- matrix(rnorm(n*p), n, p)
U[Omega] <- X[Omega]
lambda <- 1
pg_sol <- prox_grad(X, U, Omega, lambda,1)
}
