% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completion_sol.R
\name{prox_nuc_sol}
\alias{prox_nuc_sol}
\title{Proximal mapping of the scaled nuclear norm}
\usage{
prox_nuc_sol(X, gamma)
}
\arguments{
\item{X}{input matrix}

\item{gamma}{scale parameter}
}
\description{
Proximal mapping of the scaled nuclear norm
}
\examples{
set.seed(12345)
n <- 10; p <- 20
X <- matrix(rnorm(n*p), n, p)
Y <- prox_nuc_sol(X, 1)
}
