% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(y, X, lambda)
}
\arguments{
\item{y}{response variables}

\item{X}{design matrix}

\item{lambda}{vector of tuning parameters}
}
\description{
\code{ridge_regression} returns the ridge regression coefficient estimates
for a sequence of regularization parameter values.
}
