% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completion_sol.R
\name{soft_impute_sol}
\alias{soft_impute_sol}
\title{Soft-Impute Algorithm}
\usage{
soft_impute_sol(X, U, Omega, lambda, max_iter = 100, tol = 0.001)
}
\arguments{
\item{X}{input matrix}

\item{U}{current guess}

\item{Omega}{index set}

\item{lambda}{regularization parameter}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Soft-Impute Algorithm
}
\examples{
set.seed(12345)
n <- 10; p <- 20
X <- matrix(rnorm(n*p), n, p)
numel <- n*p
Omega <- sample(1:numel, floor(0.5*numel), replace=FALSE)
U <- matrix(rnorm(n*p), n, p)
U[Omega] <- X[Omega]
lambda <- 1
si_sol <- soft_impute_sol(X, U, Omega, lambda)
}
