% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{softthreshold}
\alias{softthreshold}
\title{Softthreshold operator}
\usage{
softthreshold(x, lambda)
}
\arguments{
\item{x}{input vector}

\item{lambda}{threshold}
}
\description{
Softthreshold operator
}
\examples{
softthreshold(seq(-10, 10, length.out=10), 1)
}
