% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completion_sol.R
\name{softthreshold_sol}
\alias{softthreshold_sol}
\title{Softthreshold operator}
\usage{
softthreshold_sol(x, lambda)
}
\arguments{
\item{x}{input vector}

\item{lambda}{threshold}
}
\description{
Softthreshold operator
}
\examples{
softthreshold_sol(seq(-10, 10, length.out=10), 1)
}
