% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_k.R
\name{sweep_k}
\alias{sweep_k}
\title{Sweep k}
\usage{
sweep_k(A, k)
}
\arguments{
\item{A}{The input symmetric matrix}

\item{k}{Diagonal index on which to sweep}
}
\description{
\code{sweep_k} applies the sweep operator to a symmetric matrix  on the kth diagonal entry if it is possible.
}
