% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework5_sol.R
\name{Xtilde}
\alias{Xtilde}
\title{Make reduced design matrix}
\usage{
Xtilde(y, X, beta, delta)
}
\arguments{
\item{y}{binary response}

\item{X}{design matrix}

\item{beta}{regression coefficient vector}

\item{delta}{shape parameter}
}
\description{
Make reduced design matrix
}
